Structural Changes to the code base
We are basing our code off the highschool ranker code. I got rid of the things we dont need and the things we can use. The physics function file will be used for the functions and the streamlit file will be used for the website that will be hosted on stream lit. I can work on streamlit while you (val) can work on the functons.